requirements.txt
amqp==2.6.0
asgiref==3.2.10
autopep8==1.5.3
billiard==3.6.3.0
boto3==1.14.9
botocore==1.17.9
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.4.0
django-filter==2.3.0
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
drf-yasg==1.17.1
flower==0.9.4
future==0.18.2
gunicorn==20.0.4
humanize==0.5.1
idna==2.9
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==4.6.11
Markdown==3.2.2
MarkupSafe==1.1.1
packaging==20.4
psycopg2-binary==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
python-http-client==3.2.7
pytz==2020.1
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
sendgrid==6.3.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
tornado==6.0.4
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.1.0
email-validator==1.1.1